Revert "refactor: use new deliverContent api (#36)"
This reverts commit c9eb975287d0480cf4132c14399bd0e73ffe0fcd.
