+++ src/events/newArticle.ts
@@ -1,8 +1,9 @@
-import { Client } from "discord.js";
+import debounce from "debounce";
+import { Client, Role } from "discord.js";
 import { Item } from "rss-parser";
 
+import { deliverContent } from "../utils/bot.js";
 import { logger } from "../utils/logging.js";
-import { deliverContentToAll } from "../utils/messaging.js";
 
 export type HltvArticle = Item & {
   pubDate: Date;
@@ -13,40 +14,127 @@ export type HltvArticle = Item & {
   };
 };
 
+type StatsType = {
+  server: {
+    count: number;
+    members: number;
+    withChannel: {
+      count: number;
+      members: number;
+      withRole: {
+        count: number;
+        members: number;
+      };
+    };
+  };
+  message: {
+    errors: { [key: string]: number };
+    members: number;
+    roles: number;
+  };
+};
+
+const logStats = debounce((guid, stats) => {
+  logger.info(`Article ${guid} sent to guilds`, { stats });
+}, 10e3);
+
 export default function newArticle(client: Client, article: HltvArticle) {
   logger.info("New article received", article);
 
-  const message = {
-    content: `${article.title} ${article.link}`,
-    embeds: [
-      {
-        title: article.title,
-        description: article.content,
-        url: article.link,
-        color: 0x3c6ea1,
-        author: {
-          name: "HLTV",
-          icon_url:
-            "https://www.hltv.org/img/static/favicon/apple-touch-icon.png",
-          url: "https://www.hltv.org",
+  const stats: StatsType = {
+    server: {
+      count: 0,
+      members: 0,
+      withChannel: {
+        count: 0,
+        members: 0,
+        withRole: {
+          count: 0,
+          members: 0,
         },
-        image: {
-          url: article.media.$.url,
-        },
-        footer: {
-          text: "HLTV.org - The home of competitive Counter-Strike",
-          icon_url:
-            "https://www.hltv.org/img/static/favicon/apple-touch-icon.png",
-        },
-        timestamp: article.isoDate,
       },
-    ],
+    },
+    message: {
+      errors: {},
+      members: 0,
+      roles: 0,
+    },
   };
 
-  deliverContentToAll(
-    client,
-    article.title ?? "[HLTV News Story]",
-    message,
-    article.guid,
-  );
+  let channel;
+  let role: Role | undefined;
+  let message;
+  client.guilds.cache.forEach((guild) => {
+    stats.server.count++;
+    stats.server.members += guild.memberCount;
+
+    channel = guild.channels.cache.find(
+      (channel) => channel.name === "news-feed",
+    );
+
+    if (
+      !channel ||
+      (channel.type !== "GUILD_TEXT" && channel.type !== "GUILD_FORUM")
+    ) {
+      return;
+    }
+
+    stats.server.withChannel.count++;
+    stats.server.withChannel.members += guild.memberCount;
+
+    role = guild.roles.cache.find((role) => role.name === "hltv");
+
+    message = {
+      content: `${article.title} ${article.link}`,
+      embeds: [
+        {
+          title: article.title,
+          description: article.content,
+          url: article.link,
+          color: 0x3c6ea1,
+          author: {
+            name: "HLTV",
+            icon_url:
+              "https://www.hltv.org/img/static/favicon/apple-touch-icon.png",
+            url: "https://www.hltv.org",
+          },
+          image: {
+            url: article.media.$.url,
+          },
+          footer: {
+            text: "HLTV.org - The home of competitive Counter-Strike",
+            icon_url:
+              "https://www.hltv.org/img/static/favicon/apple-touch-icon.png",
+          },
+          timestamp: article.isoDate,
+        },
+      ],
+    };
+
+    if (role) {
+      message.content = message.content.concat(` <@&${role.id}>`);
+      stats.server.withChannel.withRole.count++;
+      stats.server.withChannel.withRole.members += guild.memberCount;
+    }
+
+    let errored: boolean;
+    deliverContent(channel, article.title ?? "[HLTV News Story]", message)
+      .catch((error: Error) => {
+        errored = true;
+
+        if (!(error.message in stats.message.errors)) {
+          stats.message.errors[error.message] = 0;
+        }
+        stats.message.errors[error.message]++;
+      })
+      .finally(() => {
+        if (!errored) {
+          stats.message.members += guild.memberCount;
+
+          if (role) stats.message.roles++;
+        }
+
+        logStats(article.guid, stats);
+      });
+  });
 }
