+++ src/events/ready.ts
@@ -1,8 +1,7 @@
 import { Client } from "discord.js";
 
-import { setCommands, updateActivity } from "../utils/bot.js";
+import { rssChecker, setCommands, updateActivity } from "../utils/bot.js";
 import { logger } from "../utils/logging.js";
-import { rssChecker } from "../utils/rss.js";
 
 export default function ready(client: Client) {
   logger.info("Online");
