+++ src/utils/bot.ts
@@ -1,15 +1,32 @@
 import { SlashCommandBuilder } from "@discordjs/builders";
 import { REST } from "@discordjs/rest";
-import { ActivitiesOptions, Client, Collection } from "discord.js";
+import {
+  ActivitiesOptions,
+  Client,
+  Collection,
+  ForumChannel,
+  GuildForumThreadMessageCreateOptions,
+  TextChannel,
+  ThreadAutoArchiveDuration,
+} from "discord.js";
 import { Routes } from "discord-api-types/v9";
-import { readFileSync } from "fs";
+import { readFileSync, writeFileSync } from "fs";
 import { join } from "path";
+import Parser, { Output } from "rss-parser";
 
 import help from "../commands/help.js";
 import invite from "../commands/invite.js";
 import mute from "../commands/mute.js";
 import notify from "../commands/notify.js";
+import { HltvArticle } from "../events/newArticle.js";
 import { logger } from "../utils/logging.js";
+
+const rss = new Parser({
+  customFields: {
+    item: ["pubDate", ["media:content", "media", { keepArray: false }]],
+  },
+  timeout: 5000,
+});
 
 const liveEventsLocation = join(
   __dirname,
@@ -37,6 +54,58 @@ export function updateActivity(client: Client) {
   const random = Math.floor(Math.random() * userActivities.length);
   client.user?.setPresence({
     activities: [userActivities[random]],
+  });
+}
+
+export const deliverContent = (
+  channel: TextChannel | ForumChannel,
+  name: string,
+  message: GuildForumThreadMessageCreateOptions,
+  autoArchiveDuration: ThreadAutoArchiveDuration = 60,
+) => {
+  if (channel.type === "GUILD_FORUM") {
+    return channel.threads.create({
+      name,
+      autoArchiveDuration,
+      message,
+    });
+  }
+
+  return channel.send(message);
+};
+
+export function postUpdate(
+  client: Client,
+  content: string,
+  title: string,
+  description: string,
+) {
+  const message = {
+    content,
+    embeds: [
+      {
+        title,
+        description,
+      },
+    ],
+  };
+
+  logger.info("Posting update to servers", message);
+
+  let channel;
+  client.guilds.cache.forEach((guild) => {
+    channel = guild.channels.cache.find(
+      (channel) => channel.name === "news-feed",
+    );
+
+    if (
+      !channel ||
+      (channel.type !== "GUILD_TEXT" && channel.type !== "GUILD_FORUM")
+    ) {
+      return;
+    }
+
+    deliverContent(channel, title, message).catch(() => {});
   });
 }
 
@@ -71,4 +140,42 @@ export async function declareSlashCommands(commands: SlashCommandBuilder[]) {
   } catch (error) {
     logger.error("Error declaring slash commands", error);
   }
+}
+
+export function rssChecker(name: string, url: string, client: Client) {
+  const articleStorageFileLocation = join(
+    __dirname,
+    "..",
+    "..",
+    "storage",
+    `current_${name}_article.json`,
+  );
+
+  (async () => {
+    await rss.parseURL(url, function (error: Error, feed: Output<HltvArticle>) {
+      if (error) {
+        logger.error(`Error processing RSS feed ${url}`, error);
+        return;
+      }
+
+      const newestArticle = feed.items[0];
+
+      const file = readFileSync(articleStorageFileLocation);
+      const currentArticle = JSON.parse(file.toString());
+
+      const currentArticleDate = new Date(currentArticle.pubDate);
+      const newestArticleDate = new Date(newestArticle.pubDate);
+      const isStale = newestArticleDate < currentArticleDate;
+
+      if (
+        currentArticle.guid &&
+        newestArticle.guid !== currentArticle.guid &&
+        !isStale
+      ) {
+        const data = JSON.stringify(newestArticle);
+        writeFileSync(articleStorageFileLocation, data);
+        client.emit("newArticle", newestArticle);
+      }
+    });
+  })();
 }
